Crate: move clip-path mask rasterization to rendering
`BasicShape::render_mask` (path building + `render_mask` rasterization)
moves to `rendering::canvas::mask` as `render_clip_shape_mask`, with
`resolve_radius` and `From<FillRule> for Fill`. The shape value types and
their `FromCss`/`ToCss` stay in `style`; `clip_path.rs` no longer depends
on `rendering`.
